[mqtt] Fix Certificate/Public-Key Pinning Storage
On initial connection to a SSL enabled broker the add-on is learning
the certificate digest hash and/or the public key hash and stores them
in the thing config (pinning) for later verification of the connection.
The format of the stored hash is "method:hash" (where method is the
digest algorithm, such as SHA-256). The current code does not store the
"method" prefix, which causes subsequent use of the learned pins to
throw "Algorithm is missing" exception. This commit fixes it by:
- Adding a getMethod() public method to the Pin class that exposes
  the algorithm this pin is using.
- Modifying the pinnedLearnedHash() to store the hash in the proper
  format.